Split plotting into two tab types
